requirement.txt
fastapi
mysql-connector-python
pydantic
pydantic-extra-types
pyjwt
sqlalchemy
uvicorn
pytest